requirements.txt
istari-digital-client>=10.6
python-dotenv>=1.0.1
pandas>=2.2.0
tqdm>=4.66.0
